\name{mdrdata}
\docType{data}
\alias{mdrdata}
\title{sample data}
\description{

}
\usage{data(mdrdata)}
\format{A matrix containing 401 rows and 21 columns.}
\references{
Hahn LW, Ritchie MD, Moore JH.: Multifactor dimensionality reduction software for detecting gene-gene and gene-environment 
interactions.Bioinformatics. 2003 Feb 12;19(3):376-82 
}
\keyword{datasets}